идея лагает. Лучше засейвлю
