--- src/main/java/edu/project4/ImageProcessor.java
@@ -1,9 +1,0 @@
-package edu.project4;
-
-@FunctionalInterface
-public
-interface ImageProcessor {
-
-    void process(FractalImage image);
-
-}
