--- src/main/java/edu/project4/ImageUtils.java
@@ -1,10 +1,0 @@
-package edu.project4;
-
-public final class ImageUtils {
-    public static final double X_MAX =  0.98722222222;
-    public static final double X_MIN =  -0.98722222222;
-    public static final double Y_MAX =  0.6;
-    public static final double Y_MIN =  -0.6;
-
-
-}
