+++ src/main/java/edu/project4/README.md
@@ -1,0 +1,3 @@
+Генератор фракталов. Руководство.
+
+В аргументах при запсуке указывайте только папку с расположением .properties файла.
