--- src/main/java/edu/project4/Transformation.java
@@ -1,9 +1,0 @@
-package edu.project4;
-
-import edu.project4.components.Point;
-import java.util.function.Function;
-
-public interface Transformation extends Function<Point, Point> {
-
-}
-
