+++ src/main/java/edu/project4/image/ImageProcessor.java
@@ -1,0 +1,9 @@
+package edu.project4.image;
+
+@FunctionalInterface
+public
+interface ImageProcessor {
+
+    void process(FractalImage image);
+
+}
