+++ src/main/java/edu/project4/render/Renderer.java
@@ -1,9 +1,8 @@
 package edu.project4.render;
 
-import edu.project4.FractalImage;
+import edu.project4.image.FractalImage;
 
 public interface Renderer {
     FractalImage makeImage();
-
 
 }
