+++ src/main/java/edu/project4/systeminteraction/.properties
@@ -1,0 +1,7 @@
+fileName=myImage
+outputFolder=src/main/resources
+fileExtension=jpg
+width=800
+height=600
+symmetry=1
+optimize=true
