+++ src/main/java/edu/project4/systeminteraction/ImageProperties.java
@@ -1,0 +1,14 @@
+package edu.project4.systeminteraction;
+
+public record ImageProperties (
+        String fileName,
+        String outputFolder,
+        String fileExtension,
+        int width,
+        int height,
+        boolean optimize,
+        int symmetry
+    )
+{
+
+}
