+++ src/main/java/edu/project4/systeminteraction/SystemUtils.java
@@ -1,0 +1,15 @@
+package edu.project4.systeminteraction;
+
+import edu.project4.image.ImageFormat;
+
+public class SystemUtils {
+    public static final String DEFAULT_OUTPUT_FOLDER = "src/main/resources/";
+    public static final String DEFAULT_PROPERTIES_FOLDER = "edu/project4/systeminteraction/";
+    public static final String PROPERTIES_FILENAME = ".properties";
+    public static final ImageFormat DEFAULT_IMAGE_FORMAT = ImageFormat.PNG;
+    public static final int DEFAULT_WIDTH = 2160;
+    public static final int DEFAULT_HEIGHT = 1440;
+    public static final Boolean DEFAULT_OPTIMIZATION = true;
+    public static final int LOWER_BOUND_OF_SYMMETRY = 1;
+    public static final int UPPER_BOUND_OF_SYMMETRY = 13;
+}
