+++ src/main/java/edu/project4/transformation/Transformation.java
@@ -1,0 +1,9 @@
+package edu.project4.transformation;
+
+import edu.project4.components.Point;
+import java.util.function.Function;
+
+public interface Transformation extends Function<Point, Point> {
+
+}
+
