+++ src/main/java/edu/project4/transformation/nonlinear/variations/SphericalTransformation.java
@@ -1,0 +1,15 @@
+package edu.project4.transformation.nonlinear.variations;
+
+import edu.project4.transformation.Transformation;
+import edu.project4.components.Point;
+
+public class SphericalTransformation implements Transformation {
+
+    public Point apply(Point point) {
+        double x = point.x() / (point.x() * point.x() + point.y() * point.y());
+        double y = point.y() / (point.x() * point.x() + point.y() * point.y());
+        return new Point(x, y);
+    }
+
+
+}
